refactor(core): move tokenizer_protocol to core; update imports and cosmic-ray excludes
