+++ ml_playground/data_pipeline/preparer.py
@@ -18,7 +18,7 @@ from ml_playground.data_pipeline.transforms.tokenization import (
 )
 from ml_playground.core.error_handling import DataError
 from ml_playground.core.tokenizer import create_tokenizer
-from ml_playground.tokenizer_protocol import Tokenizer
+from ml_playground.core.tokenizer_protocol import Tokenizer
 
 
 @dataclass(frozen=True)
