+++ ml_playground/data_pipeline/transforms/io.py
@@ -13,7 +13,7 @@ from ml_playground.core.error_handling import DataError
 from ml_playground.core.logging_protocol import LoggerLike
 from ml_playground.data_pipeline.transforms.tokenization import coerce_tokenizer_type
 from ml_playground.core.tokenizer import create_tokenizer
-from ml_playground.tokenizer_protocol import Tokenizer
+from ml_playground.core.tokenizer_protocol import Tokenizer
 
 __all__ = [
     "write_bin_and_meta",
