+++ pyproject.toml
@@ -215,7 +215,7 @@ directory = ".cache/coverage/htmlcov"
 module-path = "ml_playground/checkpoint.py"
 timeout = 3.0
 excluded-modules = [
-  "ml_playground/tokenizer_protocol.py",
+  "ml_playground/core/tokenizer_protocol.py",
   "ml_playground/lr_scheduler.py",
 ]
 test-command = "pytest -q -n auto tests/unit/training/checkpointing/test_service.py"
